lab/tts/chatterBox/requirements.txt
numpy
chatterbox-tts
torchaudio
